refactor: exception 이름 변경 및 구조 수정
