+++ src/main/java/timetogeter/global/interceptor/response/error/GlobalExceptionHandler.java
@@ -7,7 +7,8 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 import timetogeter.global.interceptor.response.error.dto.ErrorResponse;
-import timetogeter.global.interceptor.response.error.errorbase.ErrorCode;
+import timetogeter.global.interceptor.response.error.status.BaseErrorCode;
+import timetogeter.global.interceptor.response.error.status.StatusCode;
 
 @Slf4j
 @RestControllerAdvice
@@ -32,39 +33,39 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
     @ExceptionHandler(IllegalArgumentException.class)
     public ResponseEntity<?> handleIllegalArgument(IllegalArgumentException e) {
         log.warn("handleIllegalArgument", e);
-        ErrorCode errorCode = GlobalErrorCode.INVALID_PARAMETER;
+        StatusCode errorCode = BaseErrorCode.INVALID_PARAMETER;
         return handleExceptionInternal(errorCode, e.getMessage());
     }
 
     @ExceptionHandler(Exception.class)
     public ResponseEntity<?> handleAllException(Exception ex) {
         log.warn("handleAllException", ex);
-        ErrorCode errorCode = GlobalErrorCode.INTERNAL_SERVER_ERROR;
+        StatusCode errorCode = BaseErrorCode.INTERNAL_SERVER_ERROR;
         return handleExceptionInternal(errorCode);
     }
 
-    private ResponseEntity<?> handleExceptionInternal(ErrorCode errorCode) {
+    private ResponseEntity<?> handleExceptionInternal(StatusCode errorCode) {
         return ResponseEntity.status(errorCode.getHttpStatus())
                 .body(makeErrorResponse(errorCode));
     }
 
-    private ErrorResponse makeErrorResponse(ErrorCode errorCode) {
+    private ErrorResponse makeErrorResponse(StatusCode errorCode) {
         return ErrorResponse.builder()
                 .code(errorCode.name())
-                .codenum(errorCode.getCodenum())
+                .codenum(errorCode.getCode())
                 .message(errorCode.getMessage())
                 .build();
     }
 
-    private ResponseEntity<?> handleExceptionInternal(ErrorCode errorCode, String message) {
+    private ResponseEntity<?> handleExceptionInternal(StatusCode errorCode, String message) {
         return ResponseEntity.status(errorCode.getHttpStatus())
                 .body(makeErrorResponse(errorCode, message));
     }
 
-    private ErrorResponse makeErrorResponse(ErrorCode errorCode, String message) {
+    private ErrorResponse makeErrorResponse(StatusCode errorCode, String message) {
         return ErrorResponse.builder()
                 .code(errorCode.name())
-                .codenum(errorCode.getCodenum())
+                .codenum(errorCode.getCode())
                 .message(message)
                 .build();
     }
