+++ src/main/java/timetogeter/global/interceptor/response/error/status/BaseErrorCode.java
@@ -1,13 +1,12 @@
-package timetogeter.global.interceptor.response.error;
+package timetogeter.global.interceptor.response.error.status;
 
 import lombok.*;
 import org.springframework.http.HttpStatus;
-import timetogeter.global.interceptor.response.error.errorbase.ErrorCode;
 
 //필요한 에러를 //[에러이름] 에러 란에 만들기
 @Getter
 @RequiredArgsConstructor
-public enum GlobalErrorCode implements ErrorCode {
+public enum BaseErrorCode implements StatusCode {
     //공통 에러
     INVALID_PARAMETER(1000,HttpStatus.BAD_REQUEST, "파라미터 값이 없습니다!"),
     INTERNAL_SERVER_ERROR(1001,HttpStatus.INTERNAL_SERVER_ERROR, "서버 내부 오류입니다!"),
@@ -19,7 +18,7 @@ public enum GlobalErrorCode implements ErrorCode {
 
 
     //===================================
-    private final int codenum;
+    private final int code;
     private final HttpStatus httpStatus;
     private final String message;
 }
